% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertIncludeGraphics.R
\name{insertIncludeGraphics}
\alias{insertIncludeGraphics}
\title{Paste image to file and insert \code{include_graphics} code block}
\usage{
insertIncludeGraphics()
}
\description{
Call this function as an addin to paste an image to a file and insert a code
block with \code{include_graphics}.
}
