% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMarkdownImage.R
\name{insertMarkdownImage}
\alias{insertMarkdownImage}
\title{Paste image to file and insert with markdown syntax}
\usage{
insertMarkdownImage()
}
\description{
Call this function as an addin to paste an image to a file and insert markdown
}
