% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteImgMacAddin-package.R
\docType{package}
\name{pasteImgMacAddin}
\alias{pasteImgMacAddin}
\alias{pasteImgMacAddin-package}
\title{pasteImgMacAddin}
\description{
An RStudio addin for Macs to paste an image from the clipboard into an RStudio document.
}
\details{
You need to have \code{pngpaste} installed (Homebrew will do it)
}
